# cache http probe request object