@Taoup 2024-07-24
Update: parallel benchmark result
```
root ➜ /workspaces/kubernetes (master) $ benchstat old.txt new.txt 
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/kubelet/prober
cpu: Intel(R) Xeon(R) Gold 5220R CPU @ 2.20GHz
        │    old.txt    │               new.txt                │
        │    sec/op     │    sec/op     vs base                │
Probe-8   173.10m ± 19%   21.70m ± 19%  -87.46% (p=0.000 n=10)

        │    old.txt    │               new.txt                │
        │     B/op      │     B/op      vs base                │
Probe-8   251.30Mi ± 0%   33.66Mi ± 0%  -86.61% (p=0.000 n=10)

        │   old.txt    │               new.txt               │
        │  allocs/op   │  allocs/op   vs base                │
Probe-8   3367.5k ± 0%   594.7k ± 0%  -82.34% (p=0.000 n=10)
```

the code doing this benchmark:
https://github.com/Taoup/kubernetes/commit/853845791b5f8ab1ce36ff113f9e433f2ec4c343