@aojea 2024-07-26
> update benchmark result:
> 
> ```
> benchstat old.txt new.txt 
> goos: linux
> goarch: amd64
> pkg: k8s.io/kubernetes/pkg/kubelet/prober
> cpu: Intel(R) Xeon(R) Gold 5220R CPU @ 2.20GHz
>         │   old.txt   │               new.txt               │
>         │   sec/op    │   sec/op     vs base                │
> Probe-8   557.6m ± 5%   164.7m ± 6%  -70.47% (p=0.000 n=10)
> 
>         │   old.txt    │               new.txt                │
>         │     B/op     │     B/op      vs base                │
> Probe-8   359.9Mi ± 0%   142.2Mi ± 0%  -60.49% (p=0.000 n=10)
> 
>         │   old.txt    │               new.txt               │
>         │  allocs/op   │  allocs/op   vs base                │
> Probe-8   3466.0k ± 0%   693.2k ± 0%  -80.00% (p=0.000 n=10)
> ```

can you add the benchmark to the code and reflect the results in a comment?