@Taoup 2024-07-26
> > the code doing this benchmark: [4c0578c](https://github.com/kubernetes/kubernetes/commit/4c0578c9c8e19ee1d393052b3d573e754f3a1ac0)
> 
> yeah, but add those results to the commit message and/or as a comment on top of the benchmark function, so we have a place where we can look in the code

sry I misunderstood your intent, updated the commit.